% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialValues.R
\name{initialValues}
\alias{initialValues}
\title{Initial values for estimation}
\usage{
initialValues(U, family = "Clayton")
}
\arguments{
\item{U}{(nx2) data matrix of pseudo-observations.}

\item{family}{'Gaussian' , 't' , 'Clayton' , 'Frank' , 'Gumbel'.}
}
\value{
\item{paraml}{Initial  values for the non-centrality parameters and Kendall's tau to be
included in the EstNCSCop function.}
}
\description{
This function computes initial values of non-centrality parameters and Kendall's tau
at selected points for the estimation non-central squared copula parameters. The results are not satisfactory. Do not use.
}
\examples{
param <- c(0.8, 2.5, 0.7) ;
U <- SimNCSCop('Clayton', 250, param)
param = initialValues(U, 'Clayton');


}
\author{
Bouchra R. Nasri, August 14, 2019
}
