% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallTau.R
\name{KendallTau}
\alias{KendallTau}
\title{Kendall's tau of a copula}
\usage{
KendallTau(family, alpha)
}
\arguments{
\item{family}{"Gaussian" , "t" , "Clayton" , "Frank" , "Gumbel"}

\item{alpha}{unconstrainted parameters of the copula family}
}
\value{
\item{tau}{estimated Kendall's tau}

\item{theta}{estimated copula parameter (constrained)}
}
\description{
This function computes the Kendall's tau of a copula family for a given a unconstrainted parameter alpha.
}
\examples{
KendallTau('Clayton',0)

}
\author{
Bouchra R. Nasri, August 14, 2019
}
