% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamCop.R
\name{ParamCop}
\alias{ParamCop}
\title{Gives the parameters of the copula family}
\usage{
ParamCop(family, alpha)
}
\arguments{
\item{family}{"Gaussian" , "t" , "Clayton" , "Frank" , "Gumbel"}

\item{alpha}{unconstrainted parameters of the copula family}
}
\value{
\item{theta}{Bivariate vector of constrained copula family parameters}
}
\description{
This function computes the parameter of the copula according to CRAN copula package where corresponding to the unconstrainted parameters alpha.
}
\examples{
ParamCop('Clayton',0)

}
\author{
Bouchra R. Nasri, August 14, 2019
}
