% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamTau.R
\name{ParamTau}
\alias{ParamTau}
\title{Unconstrained parameters}
\usage{
ParamTau(family, tau)
}
\arguments{
\item{family}{'Gaussian' , 't' , 'Clayton' , 'Frank' , 'Gumbel'}

\item{tau}{Kendall's tau of the copula family}
}
\value{
\item{alpha}{Unconstrainted parameter}
}
\description{
This function computes the  unconstrainted parameter alpha for a given Kendall's tau
}
\examples{
ParamTau('Clayton',0.5)

}
\author{
Bouchra R. Nasri, August 14, 2019
}
